Add HTML to the list of files to watch for changes
For some reason nodemon wasn’t restarting the server when html files
changed. Presume this is a bug that’s been introduced (the default
install of Prototype Kit does this without this change).

This commit explicitly adds .html files to the list of files for nodemon
to watch.

CHS xx